pyproject.toml
[tool.poetry]
name = "vta-backend"
version = "0.1.0"
description = "Virtual Teaching Assistant Backend"
authors = ["Julia Kotovich"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.8.1"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
Django = "5.0.4"
django-cors-headers = "4.3.1"
django-rest-framework = "0.1.0"
djangorestframework = "3.15.1"
filelock = "3.14.0"
fsspec = "2024.3.1"
huggingface-hub = "0.23.0"
idna = "3.7"
Jinja2 = "3.1.3"
joblib = "1.4.2"
MarkupSafe = "2.1.5"
mpmath = "1.3.0"
networkx = "3.3"
numpy = "1.26.4"
packaging = "24.0"
PyYAML = "6.0.1"
regex = "2024.4.28"
requests = "2.31.0"
safetensors = "0.4.3"
scikit-learn = "1.4.2"
scipy = "1.13.0"
sqlparse = "0.5.0"
sympy = "1.12"
threadpoolctl = "3.5.0"
tokenizers = "0.19.1"
torch = "2.3.0"
tqdm = "4.66.4"
transformers = "4.40.1"
typing_extensions = "4.11.0"
urllib3 = "2.2.1"
google = "^3.0.0"
nltk = "^3.9.1"
llama-cpp-python = "^0.3.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
